requirements.txt
fastapi
pydantic-settings
sqlmodel
asyncpg
redis
sqlalchemy
alembic
uvicorn
pytest
psycopg2-binary